+++ views/pages/page_tour/page-content.php
@@ -5,22 +5,25 @@
     <section class="tour">
         <h3 class="tour-title">Concerts</h3>
         <div class="tour-body">
+
             <div class="tour-event">
-                <time class="tour-event-date">23 Juillet 2021 - à 21h</time>
+                <time class="tour-event-date">21 Août 2021</time>
                 <p class="tour-event-location">
-                    en 1ère partie, au 'Paradiso'
-                    <a class="tour-event-mapIcon" href="https://goo.gl/maps/tmsSc3F3c1C9cQRd7" target="_blank"><i class="fas fa-map"></i></a>
+                    Concert acoustique, <br>Librairie "De Plume &amp; d'Épée"
+                    <a class="tour-event-mapIcon" href="https://goo.gl/maps/fEder5hRUjXE7ghR9" target="_blank"><i class="fas fa-map"></i></a>
                 </p>
-                <p class="tour-event-city">St Martin-En-Haute (Rhône)</p>
+                <p class="tour-event-city">Crémieu (Isère)</p>
             </div>
+            
             <div class="tour-event">
-                <time class="tour-event-date">07 Août 2021</time>
+                <time class="tour-event-date">11 Septembre 2021</time>
                 <p class="tour-event-location">
-                    en 1ère partie, à l' Espace Culturel 'Jean Carmet'
-                    <a class="tour-event-mapIcon" href="https://goo.gl/maps/m6fNwUnZwG9DxDAaA" target="_blank"><i class="fas fa-map"></i></a>
+                    Concert privé sur invitation:<br>double live, en collaboration avec le groupe "Rocking Chairs"
+                    <!-- <a class="tour-event-mapIcon" href="https://goo.gl/maps/m6fNwUnZwG9DxDAaA" target="_blank"><i class="fas fa-map"></i></a> -->
                 </p>
-                <p class="tour-event-city">Mornant (Rhône)</p>
+                <p class="tour-event-city">Beynost (Ain)</p>
             </div>
+            
             <div class="tour-event">
                 <time class="tour-event-date">06 Novembre 2021 - à 20h</time>
                 <p class="tour-event-location">
@@ -29,6 +32,7 @@
                 </p>
                 <p class="tour-event-city">Ceyssat (Puis De Dôme)</p>
             </div>
+
         </div>
     </section>
 </main>